requirements.txt
requests==2.10.0
RPi.GPIO==0.6.2
